+++ src/Common/Constants.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the Location bundle package (@see https://github.com/apacheborys/location-bundle).
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ * @license    MIT License
+ */
+
+namespace ApacheBorys\Location;
+
+/**
+ * @author Borys Yermokhin <borys_ermokhin@yahoo.com>
+ */
+final class Constants
+{
+    const EARTH_RADIUS = 6372.795477598;
+}